feat(rm): widen VM syscall ABI to intptr_t (engine side, 64-bit native modules)
The VM<->module boundary passed pointers through 32-bit int args, which
truncates on x64. Widen the syscall direction to intptr_t so native 64-bit
game modules can pass pointers across syscalls. Engine half only (both
binaries still build); the module-side (g_/cg_/ui_syscalls + vmMain/dllEntry)
and the module DLL targets follow.

- vm_local.h: systemCall -> intptr_t(*)(intptr_t*); entryPoint returns intptr_t.
- qcommon.h: VM_Create/VM_Call/VM_ArgPtr/VM_ExplicitArgPtr/Sys_LoadDll widened;
  add floatint_t union for the float-arg ABI.
- vm.c: VM_DllSyscall + VM_Call now ALWAYS marshal varargs via va_arg into an
  array (the old (&arg)[i] stack-indexing is undefined on x86_64/ARM); widen
  VM_ArgPtr/VM_ExplicitArgPtr; VM_Call returns intptr_t.
- win_main.c: Sys_LoadDll + dllEntry/vmMain proc casts widened to intptr_t.
- sv_game.c / cl_cgame.c / cl_ui.c: *SystemCalls return intptr_t, take
  intptr_t* args; VMF() now reads the float bit-pattern from the low half of
  the arg via floatint_t (was ((float*)args)[x], wrong once args are 8 bytes);
  memset/memcpy/strncpy trap returns cast (intptr_t) not (int).

Sim logic unchanged (this is ABI plumbing, not gameplay). Both etrm.exe and
etrmded.exe rebuild clean on MSVC x64.

+++ src/win32/win_main.c
@@ -572,10 +572,10 @@ char* Sys_GetDLLName( const char *name ) {
 // fqpath param added 2/15/02 by T.Ray - Sys_LoadDll is only called in vm.c at this time
 // fqpath will be empty if dll not loaded, otherwise will hold fully qualified path of dll module loaded
 // fqpath buffersize must be at least MAX_QPATH+1 bytes long
-void * QDECL Sys_LoadDll( const char *name, char *fqpath, int( QDECL **entryPoint ) ( int, ... ),
-						  int ( QDECL *systemcalls )( int, ... ) ) {
+void * QDECL Sys_LoadDll( const char *name, char *fqpath, intptr_t( QDECL **entryPoint ) ( int, ... ),
+						  intptr_t ( QDECL *systemcalls )( intptr_t, ... ) ) {
 	HINSTANCE libHandle;
-	void ( QDECL * dllEntry )( int ( QDECL *syscallptr )( int, ... ) );
+	void ( QDECL * dllEntry )( intptr_t ( QDECL *syscallptr )( intptr_t, ... ) );
 	char    *basepath;
 	char    *cdpath;
 	char    *gamedir;
@@ -625,8 +625,8 @@ void * QDECL Sys_LoadDll( const char *name, char *fqpath, int( QDECL **entryPoin
 		}
 	} else {Q_strncpyz( fqpath, fn, MAX_QPATH ) ;       // added 2/15/02 by T.Ray
 	}
-	dllEntry = ( void ( QDECL * )( int ( QDECL * )( int, ... ) ) )GetProcAddress( libHandle, "dllEntry" );
-	*entryPoint = ( int ( QDECL * )( int,... ) )GetProcAddress( libHandle, "vmMain" );
+	dllEntry = ( void ( QDECL * )( intptr_t ( QDECL * )( intptr_t, ... ) ) )GetProcAddress( libHandle, "dllEntry" );
+	*entryPoint = ( intptr_t ( QDECL * )( int,... ) )GetProcAddress( libHandle, "vmMain" );
 	if ( !*entryPoint || !dllEntry ) {
 		FreeLibrary( libHandle );
 		return NULL;
